Fix configure-info-directory corruption on Linux
Patching dumped configure-info-directory during conda install will not
update the dumped string length. This results in a string with
trailing padding zeros.

Ensure this string is stored as a zero terminated string and set the
lisp variable on startup.

Add a test that the variable contains an expected string after
installation.

OSX is not affected, because configure-info-directory does not contain
the prefix.